Enable Evas GL renderer for all profiles (#312)
For devices other than wearable, we can use either ElmFlutterView or EcoreWl2Window,
depending on apps. This commit is to avoid being device dependent.

Remove the use_evas_gl_renderer build flag and use the WEARABLE_PROFILE macro if necessary.
Before creating the engine, the embedding can select the renderer type in the engine property.
At runtime, it selects an appropriate renderer and external texture
according to the renderer type and renders to it.